+++ src/types/types.ts
@@ -97,25 +97,57 @@ export interface AlchemySettings {
  */
 export enum Network {
   ETH_MAINNET = 'eth-mainnet',
+  /** @deprecated */
   ETH_GOERLI = 'eth-goerli',
   ETH_SEPOLIA = 'eth-sepolia',
   OPT_MAINNET = 'opt-mainnet',
+  /** @deprecated */
   OPT_GOERLI = 'opt-goerli',
   OPT_SEPOLIA = 'opt-sepolia',
   ARB_MAINNET = 'arb-mainnet',
+  /** @deprecated */
   ARB_GOERLI = 'arb-goerli',
   ARB_SEPOLIA = 'arb-sepolia',
   MATIC_MAINNET = 'polygon-mainnet',
+  /** @deprecated */
   MATIC_MUMBAI = 'polygon-mumbai',
   MATIC_AMOY = 'polygon-amoy',
   ASTAR_MAINNET = 'astar-mainnet',
   POLYGONZKEVM_MAINNET = 'polygonzkevm-mainnet',
+  /** @deprecated */
   POLYGONZKEVM_TESTNET = 'polygonzkevm-testnet',
+  POLYGONZKEVM_CARDONA = 'polygonzkevm-cardona',
   BASE_MAINNET = 'base-mainnet',
   BASE_GOERLI = 'base-goerli',
   BASE_SEPOLIA = 'base-sepolia',
   ZKSYNC_MAINNET = 'zksync-mainnet',
-  ZKSYNC_SEPOLIA = 'zksync-sepolia'
+  ZKSYNC_SEPOLIA = 'zksync-sepolia',
+  SHAPE_MAINNET = 'shape-mainnet',
+  SHAPE_SEPOLIA = 'shape-sepolia',
+  LINEA_MAINNET = 'linea-mainnet',
+  LINEA_SEPOLIA = 'linea-sepolia',
+  FANTOM_MAINNET = 'fantom-mainnet',
+  FANTOM_TESTNET = 'fantom-testnet',
+  ZETACHAIN_MAINNET = 'zetachain-mainnet',
+  ZETACHAIN_TESTNET = 'zetachain-testnet',
+  ARBNOVA_MAINNET = 'arbnova-mainnet',
+  BLAST_MAINNET = 'blast-mainnet',
+  BLAST_SEPOLIA = 'blast-sepolia',
+  MANTLE_MAINNET = 'mantle-mainnet',
+  MANTLE_SEPOLIA = 'mantle-sepolia',
+  SCROLL_MAINNET = 'scroll-mainnet',
+  SCROLL_SEPOLIA = 'scroll-sepolia',
+  GNOSIS_MAINNET = 'gnosis-mainnet',
+  GNOSIS_CHIADO = 'gnosis-chiado',
+  BNB_MAINNET = 'bnb-mainnet',
+  BNB_TESTNET = 'bnb-testnet',
+  AVAX_MAINNET = 'avax-mainnet',
+  AVAX_FUJI = 'avax-fuji',
+  CELO_MAINNET = 'celo-mainnet',
+  CELO_ALFAJORES = 'celo-alfajores',
+  METIS_MAINNET = 'metis-mainnet',
+  OPBNB_MAINNET = 'opbnb-mainnet',
+  OPBNB_TESTNET = 'opbnb-testnet'
 }
 
 /** Token Types for the `getTokenBalances()` endpoint. */
@@ -1202,6 +1234,25 @@ export interface CustomGraphqlWebhookParams {
    * created on network of the app provided in the api key config.
    */
   network?: Network;
+  /**
+   * Whether to only receive webhooks if the query on the block is not empty.
+   * Defaults to false.
+   */
+  skipEmptyMessages?: boolean;
+  /**
+   * App IDs are now required for graphQL webhooks. You can find the app ID
+   * following the steps here:
+   * {@link https://docs.alchemy.com/reference/notify-api-faq#where-can-i-find-the-app-id}.
+   *
+   * The webhook will be created on the app and network associated with the appId.
+   * To find the app id of a project, go to the Alchemy Dashboard in the Apps tab.
+   * After clicking on an app, the app id is the string in the URL following 'apps/'.
+   *
+   * Note that although this property is marked as optional, it is *actually required*
+   * for creating a custom GraphQL webhook. This is a workaround to avoid a breaking
+   * change in the API.
+   */
+  appId?: string;
 }
 
 /**